update: extracted AddAddress to a separate component
